fix: fix calculating temp for mems

+++ src/rs_driver/driver/decoder/decoder_RSEOS.hpp
@@ -50,7 +50,7 @@ typedef struct
   RSTimestampUTC timestamp;
   uint8_t reserved[10];
   uint8_t lidar_type;
-  int8_t temperature;
+  uint8_t temperature;
 } RSEOSMsopHeader;
 
 typedef struct
@@ -159,7 +159,7 @@ inline bool DecoderRSEOS<T_PointCloud>::decodeMsopPkt(const uint8_t* packet, siz
   const RSEOSMsopPkt& pkt = *(RSEOSMsopPkt*)packet;
   bool ret = false;
 
-  this->temperature_ = static_cast<float>(pkt.header.temperature - this->const_param_.TEMPERATURE_RES);
+  this->temperature_ = static_cast<float>((int)pkt.header.temperature - this->const_param_.TEMPERATURE_RES);
 
   double pkt_ts = 0;
   if (this->param_.use_lidar_clock)
