+++ src/rs_driver/driver/decoder/decoder_RSM1.hpp
@@ -49,7 +49,7 @@ typedef struct
   RSTimestampUTC timestamp;
   uint8_t reserved[10];
   uint8_t lidar_type;
-  int8_t temperature;
+  uint8_t temperature;
 } RSM1MsopHeader;
 
 typedef struct
@@ -216,7 +216,7 @@ inline bool DecoderRSM1<T_PointCloud>::decodeMsopPkt(const uint8_t* packet, size
   const RSM1MsopPkt& pkt = *(RSM1MsopPkt*)packet;
   bool ret = false;
 
-  this->temperature_ = static_cast<float>(pkt.header.temperature - this->const_param_.TEMPERATURE_RES);
+  this->temperature_ = static_cast<float>((int)pkt.header.temperature - this->const_param_.TEMPERATURE_RES);
 
   double pkt_ts = 0;
   if (this->param_.use_lidar_clock)
