+++ src/rs_driver/driver/decoder/decoder_RSM2.hpp
@@ -50,7 +50,7 @@ typedef struct
   RSTimestampUTC timestamp;
   uint8_t reserved[10];
   uint8_t lidar_type;
-  int8_t temperature;
+  uint8_t temperature;
 } RSM2MsopHeader;
 
 typedef struct
@@ -162,7 +162,7 @@ inline bool DecoderRSM2<T_PointCloud>::decodeMsopPkt(const uint8_t* packet, size
   const RSM2MsopPkt& pkt = *(RSM2MsopPkt*)packet;
   bool ret = false;
 
-  this->temperature_ = static_cast<float>(pkt.header.temperature - this->const_param_.TEMPERATURE_RES);
+  this->temperature_ = static_cast<float>((int)pkt.header.temperature - this->const_param_.TEMPERATURE_RES);
 
   double pkt_ts = 0;
   if (this->param_.use_lidar_clock)
